requirements.txt
# Dependencies for the algorithm
tqdm
numpy
scipy
pandas
pytorch

# Generate figures
matplotlib
seaborn